requirements.txt
Fraction
numpy
matplotlib
opencv-python
piexif==1.1.3
PyQt5==5.13.2
PyQt5-sip==12.7.0
pyrealsense2
pyserial==3.4
tensorflow
tensorflow-gpu